Extract StatusThreadingConcern from Status
